agent_multi_friend_scheduling/pyproject.toml
[project]
name = "a2a_friend_scheduling"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-adk>=1.2.1",
  "a2a-sdk>=0.2.5",
  "nest-asyncio>=1.6.0",
  "python-dotenv",
  "click",
  "uvicorn",
  "google-generativeai",
  "httpx"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
